Changed node validation predicate
It now includes the previous node in addition to the current node. This allows for checking, for example, if the current node is on a floor is higher than the previous node while the previous node was at a height of 2 (this would occur if the enemy can jump into a stack of 2 tiles high).