Progress on informed A*
Right now, because I'm using previous information, nodes that are "invalid" can end up being added to the list of valid nodes if the agent somehow places themself so that their starting node is in an invalid position.
TODO: if using the informed A*, validate the starting node. If validation fails, search for a nearby valid node to start the pathfinding from (based on a function given by the agent), defaulting to finding the first tile standable tile below where the agent currently is.