Started creating SignalAcceptor
SignalAcceptor will be the core of a signal system used to convey (of course) signals across different data objects. The power of these signals is that they store parameters in a string which is parsed and can be optionally accepted. For example, sending a knockback signal through attackdata can be ignored by enemies that don't knockback and accepted by those that do.