Created SignalMessaging package
Sending a signal now works correctly and types are correctly read out of the BaseSignalParameter class using SignalData.ReadValue<float>(string) for example.
The current solution to storing a reference to the SignalScribe is a little bit janky. Ideally, a scriptable object would be stored in the project settings which could be used.
TODO: figure out how to store the SignalScribe scriptable object in the project settings.
TODO: figure out how to automatically generate/instantiate in the asset folder and slot the built-in types to the SignalScribe when it's first made.